# vecta/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Vecta Platform API
  version: 1.0.0
  description: |
    Life-as-a-Service for F-1 international students.
    
    **Authentication:** All `/api/v1/*` endpoints (except webhooks) require
    `Authorization: Bearer <VectaIDToken>` (RS256 JWT).
    
    **F-1 Compliance:** `POST /api/v1/mobility/rides/accept` and
    `POST /api/v1/mobility/driver/go-online` always return `403 F1_VISA_COMPLIANCE_VIOLATION`
    regardless of token — these routes are architectural dead ends.
  contact:
    email: api@vecta.io
  license:
    name: Proprietary
    url: https://vecta.io/terms
servers:
  - url: https://api.vecta.io
    description: Production
  - url: https://api.staging.vecta.io
    description: Staging
  - url: http://localhost:4000
    description: Local development

# ---------------------------------------------------------------------------
# Security schemes
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    VectaIDToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 JWT issued by Vecta identity service. Contains zero PII.
  schemas:
...
    GuaranteeTier:
      type: string
      enum: [PLATINUM, GOLD, SILVER, STANDARD, INSUFFICIENT]
    TrustScoreResponse:
      type: object
      properties:
        novaScore:         { type: integer, minimum: 300, maximum: 850 }
        novaTier:          { type: string, enum: [Building, Fair, Good, Excellent] }
        compositeScore:    { type: integer, minimum: 0, maximum: 1000 }
        guaranteeTier:     { $ref: '#/components/schemas/GuaranteeTier' }
        maxRentApproval:   { type: number, description: Max monthly rent in USD }
        depositMultiplier: { type: number, enum: [1.0, 1.5, 2.0, 2.5] }
        guaranteeMonths:   { type: integer }
        breakdown:
          type: object
          properties:
            novaExplanation:      { type: string }
            liquidityExplanation: { type: string }
            visaExplanation:      { type: string }
            identityExplanation:  { type: string }
    LetterOfCredit:
      type: object
      properties:
        id:              { type: string, format: uuid }
        status:          { type: string, enum: [pending, active, expired] }
        monthlyRent:     { type: number }
        guaranteeMonths: { type: integer }
        expiresAt:       { type: string, format: date-time }
        downloadUrl:     { type: string, format: uri, description: 60-min signed S3 URL }
...
paths:
...
  /api/v1/housing/trust-score:
    get:
      summary: Get composite Vecta trust score
      security: [{ VectaIDToken: [] }]
      tags: [Housing]
      parameters:
        - name: monthlyRent
          in: query
          schema: { type: number, default: 1500 }
        - name: leaseDuration
          in: query
          schema: { type: integer, default: 12 }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TrustScoreResponse' }
  /api/v1/housing/loc/generate:
    post:
      summary: Generate Letter of Credit PDF
      security: [{ VectaIDToken: [] }]
      tags: [Housing]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [monthlyRent]
              properties:
                monthlyRent:     { type: number, minimum: 100, maximum: 50000 }
                landlordName:    { type: string }
                propertyAddress: { type: string }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LetterOfCredit' }
        '422':
          description: Bank not connected
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:  { type: string, enum: [BANK_NOT_CONNECTED] }
                  action: { type: string, enum: [CONNECT_BANK] }

  # --- Mobility (F-1 compliance enforced) ---
...